Extract segment loading code into new SegmentParser type
